fix game search cancellation and caddy www redirect
