+++ src/frontend/DockerfileCaddyMaster
@@ -3,8 +3,10 @@ FROM golang:1.23.2-alpine3.20
 WORKDIR /var/www/html/
 ARG APP_URL
 ARG BE_URL
+ARG TLS_CONFIG
 ENV APP_URL=$APP_URL
 ENV BE_URL=$BE_URL
+ENV TLS_CONFIG=$TLS_CONFIG
 COPY . /var/www/html/
 
 RUN sed -i "s#APP_URL#${APP_URL}#g" /var/www/html/static/js/settings.js  # for dynamic config of url
