+++ src/frontend/README.md
@@ -1,8 +1,4 @@
-<<<<<<< HEAD
 ## Local development
-=======
-### Develop locally
->>>>>>> master
 
 When changing one of the files within the _/static_ folder, you can use the *ops/update_fe_docker.sh* script to inject
 the changed file into the caddy container and then simply reload tha page in the browser to receive the updated file.
